removed eslintrc sourcetype and export class
