+++ src/index.html
@@ -5,7 +5,11 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>2048</title>
     <link rel="stylesheet" href="./styles/main.scss">
-    <link rel="shortcut icon" href="./images/logo.png" type="image/x-icon">
+    <link
+      rel="shortcut icon"
+      href="./images/logo.png"
+      type="image/x-icon"
+    >
   </head>
   <body>
     <div class="container">
@@ -23,7 +27,6 @@
         <tbody>
           <tr class="field-row">
             <td class="field-cell">
-              <!-- <div class="field-cell field-cell--2">2</div> -->
             </td>
             <td class="field-cell"></td>
             <td class="field-cell"></td>
